Add timeout with extra time to confirm
The method could have been an overload of TryAcquire, but I want to
keep the library compatible with the original, hence the new method
TryAcquireWithTimeout.

The GetMillisecondsToWait method might be an overkill for this case. I
will leave it in this commit just for the sake of archiving, in the
next commit it will be more simple.