Make tests more deterministic
All of these tests are non-deterministic. But given enough time they
consistently give the same results. The big exception is the one that
tries to get the lock exactly when released. If IO is sufficiently fast
then this test will succeed, but if not, it will certainly fail due to
timing out.